Move middleware to main route
